requirements.txt
Django==1.7
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.15
dj-database-url==0.3.0
dj-static==0.0.6
django-celery==3.1.16
django-grappelli==2.6.1
django-toolbelt==0.0.1
gunicorn==19.1.1
honcho==0.5.0
kombu==3.0.23
psycopg2==2.5.4
pytz==2014.7
static3==0.5.1
wsgiref==0.1.2

boto
django-storages

djangorestframework==2.4.3

